algorithms/Cracking-The-Code-Interview/chapter-1/1-6/__tests__/string-compression.spec.ts
import stringCompression from '../string-compression';

describe('String Compression', () => {
    test('a string with no dupes should return itself', () => {
        expect(stringCompression('lol')).toBe('lol');
    });
    test('a string with one dupe set returns correct result', () => {
        expect(stringCompression('loool')).toBe('l3ol');
    });
    test('a string with two dupe sets returns correct result', () => {
        expect(stringCompression('looollll')).toBe('l3o4l');
    });
    test('capital and lower case letters are distinct', () => {
        expect(stringCompression('lOol')).toBe('lOol');
    });
});
